Optimize parinfer--goto-current-toplevel with short circuiting
The result of the expression does not depend on the order as none of
them have relevant side effects. So we can reorder them and use short
circuiting to avoid running the expensive
parinfer--in-comment-or-string-p (because it uses syntax-ppss) as much
as possible.

Preliminary testing shows something like a 2x~4x speedup for this
particular function; this will depend on the file.