fix: sometimes parinfer-readjust-paren can fail
If it fails with unbalanced quotes, it could be a problem with the
chunking logic (applying parinfer only to nearby text). This would
cause parinfer-readjust-paren to fail while
parinfer-readjust-paren-buffer succeeds.

This change adds a fallback to just try parinfer-readjust-paren-buffer
when encountering an unclosed-quote error. It also adds error
reporting to parinfer-readjust-paren-buffer, so that if there is
really an unclosed quote, the error will still come through.